backend/requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
faster-whisper==1.1.*
groq==0.15.*
python-dotenv==1.0.*
pydub==0.25.*
python-multipart==0.0.*